+++ dune/xt/common/lapacke.cc
@@ -19,16 +19,9 @@
 
 #if HAVE_MKL
 #  include <mkl.h>
-#else // HAVE_MKL
-#  if HAVE_LAPACKE
-#    include <lapacke.h>
-#  endif
-#  if HAVE_CBLAS
-#    include <cblas.h>
-#  endif
-#endif // HAVE_MKL
-
-#include <dune/common/unused.hh>
+#elif HAVE_LAPACKE
+#  include <lapacke.h>
+#endif
 
 #include <dune/xt/common/exceptions.hh>
 #include <dune/xt/common/unused.hh>
@@ -41,30 +34,9 @@
 #  define DXTC_LAPACKE_ONLY(param) DXTC_UNUSED(param)
 #endif
 
-#if HAVE_MKL || HAVE_CBLAS
-#  define DXTC_CBLAS_ONLY(param) param
-#else
-#  define DXTC_CBLAS_ONLY(param) DXTC_UNUSED(param)
-#endif
-
 namespace Dune {
 namespace XT {
 namespace Common {
-namespace Mkl {
-
-
-void exp(const int n, const double* a, double* y)
-{
-#if HAVE_MKL
-  vdExp(n, a, y);
-#else
-  for (int ii = 0; ii < n; ++ii)
-    y[ii] = std::exp(a[ii]);
-#endif
-}
-
-
-} // namespace Mkl
 namespace Lapacke {
 
 
@@ -490,8 +462,6 @@ int dtrcon(int DXTC_LAPACKE_ONLY(matrix_layout),
 }
 
 
-int LAPACKE_dtrcon(int matrix_layout, char norm, char uplo, char diag, int n, const double* a, int lda, double* rcond);
-
 int dpttrf(int DXTC_LAPACKE_ONLY(n), double* DXTC_LAPACKE_ONLY(d), double* DXTC_LAPACKE_ONLY(e))
 {
 #if HAVE_MKL || HAVE_LAPACKE
@@ -576,303 +546,6 @@ int zunmqr(int DXTC_LAPACKE_ONLY(matrix_layout),
 
 
 } // namespace Lapacke
-namespace Blas {
-
-
-/**
- * \brief If true, calling any of the other methods makes sense.
- */
-bool available()
-{
-#if HAVE_MKL || HAVE_CBLAS
-  return true;
-#else
-  return false;
-#endif
-}
-
-
-int row_major()
-{
-#if HAVE_MKL || HAVE_CBLAS
-  return CblasRowMajor;
-#else
-  DUNE_THROW(Exceptions::dependency_missing, "You are missing CBLAS or the intel mkl, check available() first!");
-  return 1;
-#endif
-}
-
-
-int col_major()
-{
-#if HAVE_MKL || HAVE_CBLAS
-  return CblasColMajor;
-#else
-  DUNE_THROW(Exceptions::dependency_missing, "You are missing CBLAS or the intel mkl, check available() first!");
-  return 1;
-#endif
-}
-
-
-int left()
-{
-#if HAVE_MKL || HAVE_CBLAS
-  return CblasLeft;
-#else
-  DUNE_THROW(Exceptions::dependency_missing, "You are missing CBLAS or the intel mkl, check available() first!");
-  return 1;
-#endif
-}
-
-
-int right()
-{
-#if HAVE_MKL || HAVE_CBLAS
-  return CblasRight;
-#else
-  DUNE_THROW(Exceptions::dependency_missing, "You are missing CBLAS or the intel mkl, check available() first!");
-  return 1;
-#endif
-}
-
-
-int upper()
-{
-#if HAVE_MKL || HAVE_CBLAS
-  return CblasUpper;
-#else
-  DUNE_THROW(Exceptions::dependency_missing, "You are missing CBLAS or the intel mkl, check available() first!");
-  return 1;
-#endif
-}
-
-
-int lower()
-{
-#if HAVE_MKL || HAVE_CBLAS
-  return CblasLower;
-#else
-  DUNE_THROW(Exceptions::dependency_missing, "You are missing CBLAS or the intel mkl, check available() first!");
-  return 1;
-#endif
-}
-
-
-int trans()
-{
-#if HAVE_MKL || HAVE_CBLAS
-  return CblasTrans;
-#else
-  DUNE_THROW(Exceptions::dependency_missing, "You are missing CBLAS or the intel mkl, check available() first!");
-  return 1;
-#endif
-}
-
-
-int no_trans()
-{
-#if HAVE_MKL || HAVE_CBLAS
-  return CblasNoTrans;
-#else
-  DUNE_THROW(Exceptions::dependency_missing, "You are missing CBLAS or the intel mkl, check available() first!");
-  return 1;
-#endif
-}
-
-
-int unit()
-{
-#if HAVE_MKL || HAVE_CBLAS
-  return CblasUnit;
-#else
-  DUNE_THROW(Exceptions::dependency_missing, "You are missing CBLAS or the intel mkl, check available() first!");
-  return 1;
-#endif
-}
-
-
-int non_unit()
-{
-#if HAVE_MKL || HAVE_CBLAS
-  return CblasNonUnit;
-#else
-  DUNE_THROW(Exceptions::dependency_missing, "You are missing CBLAS or the intel mkl, check available() first!");
-  return 1;
-#endif
-}
-
-
-void dgemv(const int DXTC_CBLAS_ONLY(layout),
-           const int DXTC_CBLAS_ONLY(trans),
-           const int DXTC_CBLAS_ONLY(m),
-           const int DXTC_CBLAS_ONLY(n),
-           const double DXTC_CBLAS_ONLY(alpha),
-           const double* DXTC_CBLAS_ONLY(a),
-           const int DXTC_CBLAS_ONLY(lda),
-           const double* DXTC_CBLAS_ONLY(x),
-           const int DXTC_CBLAS_ONLY(incx),
-           const double DXTC_CBLAS_ONLY(beta),
-           double* DXTC_CBLAS_ONLY(y),
-           const int DXTC_CBLAS_ONLY(incy))
-{
-#if HAVE_MKL || HAVE_CBLAS
-  cblas_dgemv(static_cast<CBLAS_LAYOUT>(layout),
-              static_cast<CBLAS_TRANSPOSE>(trans),
-              m,
-              n,
-              alpha,
-              a,
-              lda,
-              x,
-              incx,
-              beta,
-              y,
-              incy);
-#else
-  DUNE_THROW(Exceptions::dependency_missing, "You are missing CBLAS or the intel mkl, check available() first!");
-#endif
-}
-
-
-void dtrsm(const int DXTC_CBLAS_ONLY(layout),
-           const int DXTC_CBLAS_ONLY(side),
-           const int DXTC_CBLAS_ONLY(uplo),
-           const int DXTC_CBLAS_ONLY(transa),
-           const int DXTC_CBLAS_ONLY(diag),
-           const int DXTC_CBLAS_ONLY(m),
-           const int DXTC_CBLAS_ONLY(n),
-           const double DXTC_CBLAS_ONLY(alpha),
-           const double* DXTC_CBLAS_ONLY(a),
-           const int DXTC_CBLAS_ONLY(lda),
-           double* DXTC_CBLAS_ONLY(b),
-           const int DXTC_CBLAS_ONLY(ldb))
-{
-#if HAVE_MKL || HAVE_CBLAS
-  cblas_dtrsm(static_cast<CBLAS_LAYOUT>(layout),
-              static_cast<CBLAS_SIDE>(side),
-              static_cast<CBLAS_UPLO>(uplo),
-              static_cast<CBLAS_TRANSPOSE>(transa),
-              static_cast<CBLAS_DIAG>(diag),
-              m,
-              n,
-              alpha,
-              a,
-              lda,
-              b,
-              ldb);
-#  ifndef NDEBUG
-  for (int ii = 0; ii < m; ++ii)
-    if (std::isnan(b[ii]) || std::isinf(b[ii]))
-      DUNE_THROW(Dune::MathError, "Triangular solve using cblas_dtrsm failed!");
-#  endif
-#else
-  DUNE_THROW(Exceptions::dependency_missing, "You are missing CBLAS or the intel mkl, check available() first!");
-#endif
-}
-
-
-void dtrsv(const int DXTC_CBLAS_ONLY(layout),
-           const int DXTC_CBLAS_ONLY(uplo),
-           const int DXTC_CBLAS_ONLY(transa),
-           const int DXTC_CBLAS_ONLY(diag),
-           const int DXTC_CBLAS_ONLY(n),
-           const double* DXTC_CBLAS_ONLY(a),
-           const int DXTC_CBLAS_ONLY(lda),
-           double* DXTC_CBLAS_ONLY(x),
-           const int DXTC_CBLAS_ONLY(incx))
-{
-#if HAVE_MKL || HAVE_CBLAS
-  cblas_dtrsv(static_cast<CBLAS_LAYOUT>(layout),
-              static_cast<CBLAS_UPLO>(uplo),
-              static_cast<CBLAS_TRANSPOSE>(transa),
-              static_cast<CBLAS_DIAG>(diag),
-              n,
-              a,
-              lda,
-              x,
-              incx);
-#  ifndef NDEBUG
-  for (int ii = 0; ii < n; ++ii)
-    if (std::isnan(x[ii]) || std::isinf(x[ii]))
-      DUNE_THROW(Dune::MathError, "Triangular solve using cblas_dtrsv failed!");
-#  endif
-#else
-  DUNE_THROW(Exceptions::dependency_missing, "You are missing CBLAS or the intel mkl, check available() first!");
-#endif
-}
-
-
-void ztrsm(const int DXTC_CBLAS_ONLY(layout),
-           const int DXTC_CBLAS_ONLY(side),
-           const int DXTC_CBLAS_ONLY(uplo),
-           const int DXTC_CBLAS_ONLY(transa),
-           const int DXTC_CBLAS_ONLY(diag),
-           const int DXTC_CBLAS_ONLY(m),
-           const int DXTC_CBLAS_ONLY(n),
-           const void* DXTC_CBLAS_ONLY(alpha),
-           const void* DXTC_CBLAS_ONLY(a),
-           const int DXTC_CBLAS_ONLY(lda),
-           void* DXTC_CBLAS_ONLY(b),
-           const int DXTC_CBLAS_ONLY(ldb))
-{
-#if HAVE_MKL || HAVE_CBLAS
-  cblas_ztrsm(static_cast<CBLAS_LAYOUT>(layout),
-              static_cast<CBLAS_SIDE>(side),
-              static_cast<CBLAS_UPLO>(uplo),
-              static_cast<CBLAS_TRANSPOSE>(transa),
-              static_cast<CBLAS_DIAG>(diag),
-              m,
-              n,
-              alpha,
-              a,
-              lda,
-              b,
-              ldb);
-#  ifndef NDEBUG
-  for (int ii = 0; ii < m; ++ii)
-    if (std::isnan(std::abs(static_cast<std::complex<double>*>(b)[ii]))
-        || std::isinf(std::abs(static_cast<std::complex<double>*>(b)[ii])))
-      DUNE_THROW(Dune::MathError, "Triangular solve using cblas_ztrsm failed!");
-#  endif
-#else
-  DUNE_THROW(Exceptions::dependency_missing, "You are missing CBLAS or the intel mkl, check available() first!");
-#endif
-}
-
-
-void ztrsv(const int DXTC_CBLAS_ONLY(layout),
-           const int DXTC_CBLAS_ONLY(uplo),
-           const int DXTC_CBLAS_ONLY(transa),
-           const int DXTC_CBLAS_ONLY(diag),
-           const int DXTC_CBLAS_ONLY(n),
-           const void* DXTC_CBLAS_ONLY(a),
-           const int DXTC_CBLAS_ONLY(lda),
-           void* DXTC_CBLAS_ONLY(x),
-           const int DXTC_CBLAS_ONLY(incx))
-{
-#if HAVE_MKL || HAVE_CBLAS
-  cblas_ztrsv(static_cast<CBLAS_LAYOUT>(layout),
-              static_cast<CBLAS_UPLO>(uplo),
-              static_cast<CBLAS_TRANSPOSE>(transa),
-              static_cast<CBLAS_DIAG>(diag),
-              n,
-              a,
-              lda,
-              x,
-              incx);
-#  ifndef NDEBUG
-  for (int ii = 0; ii < n; ++ii)
-    if (std::isnan(std::abs(static_cast<std::complex<double>*>(x)[ii]))
-        || std::isinf(std::abs(static_cast<std::complex<double>*>(x)[ii])))
-      DUNE_THROW(Dune::MathError, "Triangular solve using cblas_ztrsv failed!");
-#  endif
-#else
-  DUNE_THROW(Exceptions::dependency_missing, "You are missing CBLAS or the intel mkl, check available() first!");
-#endif
-}
-
-
-} // namespace Blas
 } // namespace Common
 } // namespace XT
 } // namespace Dune
