+++ dune/xt/common/lapacke.hh
@@ -14,20 +14,11 @@
 
 #include <complex>
 
+#include <dune/xt/common/deprecated.hh>
+
 namespace Dune {
 namespace XT {
 namespace Common {
-namespace Mkl {
-
-
-/**
- * \brief Wrapper around MKL's vdExp
- * \sa    vdExp
- */
-void exp(const int n, const double* a, double* y);
-
-
-} // namespace Mkl
 namespace Lapacke {
 
 
@@ -269,6 +260,7 @@ int dsygv(int matrix_layout, int itype, char jobz, char uplo, int n, double* a, 
  */
 int dtrcon(int matrix_layout, char norm, char uplo, char diag, int n, const double* a, int lda, double* rcond);
 
+
 /**
  * \brief Wrapper around LAPACKE_dpttrf
  * \sa    LAPACKE_dpttrf
@@ -315,166 +307,6 @@ int zungqr(int matrix_layout, int m, int n, int k, std::complex<double>* a, int 
 
 
 } // namespace Lapacke
-namespace Blas {
-
-
-/**
- * \brief If true, calling any of the other methods makes sense.
- */
-bool available();
-
-
-/**
- * \brief Wrapper around CBlasRowMajor
- * \sa    CblasRowMajor
- */
-int row_major();
-
-
-/**
- * \brief Wrapper around CBlasColMajor
- * \sa    CblasColMajor
- */
-int col_major();
-
-
-/**
- * \brief Wrapper around CBlasLeft
- * \sa    CblasLeft
- */
-int left();
-
-
-/**
- * \brief Wrapper around CBlasRight
- * \sa    CblasRight
- */
-int right();
-
-
-/**
- * \brief Wrapper around CblasUpper
- * \sa    CblasUpper
- */
-int upper();
-
-
-/**
- * \brief Wrapper around CblasLower
- * \sa    CblasLower
- */
-int lower();
-
-
-/**
- * \brief Wrapper around CblasTrans
- * \sa    CblasTrans
- */
-int trans();
-
-
-/**
- * \brief Wrapper around CblasNoTrans
- * \sa    CblasNoTrans
- */
-int no_trans();
-
-
-/**
- * \brief Wrapper around CblasUnit
- * \sa    CblasUnit
- */
-int unit();
-
-
-/**
- * \brief Wrapper around CblasNonUnit
- * \sa    CblasNonUnit
- */
-int non_unit();
-
-
-/**
- * \brief Wrapper around cblas_dgemv
- * \sa    cblas_dgemv
- */
-void dgemv(const int layout,
-           const int trans,
-           const int m,
-           const int n,
-           const double alpha,
-           const double* a,
-           const int lda,
-           const double* x,
-           const int incx,
-           const double beta,
-           double* y,
-           const int incy);
-
-/**
- * \brief Wrapper around cblas_dtrsm
- * \sa    cblas_dtrsm
- */
-void dtrsm(const int layout,
-           const int side,
-           const int uplo,
-           const int transa,
-           const int diag,
-           const int m,
-           const int n,
-           const double alpha,
-           const double* a,
-           const int lda,
-           double* b,
-           const int ldb);
-
-/**
- * \brief Wrapper around cblas_dtrsv
- * \sa    cblas_dtrsv
- */
-void dtrsv(const int layout,
-           const int uplo,
-           const int transa,
-           const int diag,
-           const int n,
-           const double* a,
-           const int lda,
-           double* x,
-           const int incx);
-
-/**
- * \brief Wrapper around cblas_ztrsm
- * \sa    cblas_ztrsm
- */
-void ztrsm(const int layout,
-           const int side,
-           const int uplo,
-           const int transa,
-           const int diag,
-           const int m,
-           const int n,
-           const void* alpha,
-           const void* a,
-           const int lda,
-           void* b,
-           const int ldb);
-
-/**
- * \brief Wrapper around cblas_ztrsv
- * \sa    cblas_ztrsv
- */
-void ztrsv(const int layout,
-           const int uplo,
-           const int transa,
-           const int diag,
-           const int n,
-           const void* a,
-           const int lda,
-           void* x,
-           const int incx);
-
-
-} // namespace Blas
 } // namespace Common
 } // namespace XT
 } // namespace Dune
